InfoBubbleを表示する (#15)
* wip

* Fix: InfoWindow positioning is incorrect (confirmed: Google Maps, Mapbox, and Here. Not confirmed yet: ArcGIS)

* Info bubbles follow the dragged markers.

* update: .gitignore file

* wip: Implementing the marker drag gesture for Here SDK

* fix: linter errors

* fix: duplicated markers are added

* fix: KtLint errors

* update README.md

* The mapProvider of MapViewBase is no longer necessary

* Implement toasting logic in AppViewModel

* Create Settings class for defining default values

* rename MapView.kt to MapViewContainer.kt

* Fix: Cannot follow the map camera position for ArcGIS MapView

* Fix: InfoBubble does not follow the clicked marker on ArcGIS MapView

* Fix: onMapClick does not be involved

* Optimize code structure

* Fix: InfoBubble does not follow on ArcGIS Mapview when device is rotated

* Fix: linter errors

* Launch Google Maps app when taps on the "Get Directions" button in InfoBubble

+++ mapconductor-core/src/main/java/com/mapconductor/core/geocell/KDTree.kt
@@ -1,6 +1,6 @@
 package com.mapconductor.core.geocell
 
-import com.mapconductor.core.Offset
+import androidx.compose.ui.geometry.Offset
 import java.util.PriorityQueue
 import kotlin.math.pow
 import kotlin.math.sqrt
@@ -56,7 +56,7 @@ class KDTree(
         var newBestDist = bestDist
         if (distSq < newBestDist) {
             newBest = node.cell
-            newBestDist = distSq
+            newBestDist = distSq.toDouble()
         }
 
         val (near, far) = if (queryVal < nodeVal) node.left to node.right else node.right to node.left
@@ -70,7 +70,7 @@ class KDTree(
 
     fun nearestWithDistance(query: Offset): HexCellWithDistance? {
         val cell = nearest(query) ?: return null
-        return HexCellWithDistance(cell, distanceMeters(query, cell.centerXY))
+        return HexCellWithDistance(cell, distanceMeters(query, cell.centerXY).toDouble())
     }
 
     fun nearestKWithDistance(
@@ -89,7 +89,7 @@ class KDTree(
         queue: PriorityQueue<Pair<Double, HexCell>>,
     ) {
         if (node == null) return
-        val distSq = squaredDistance(query, node.cell.centerXY)
+        val distSq = squaredDistance(query, node.cell.centerXY).toDouble()
         if (queue.size < k) {
             queue.offer(distSq to node.cell)
         } else if (distSq < queue.peek()!!.first) {
@@ -126,7 +126,7 @@ class KDTree(
         if (node == null) return
         val distSq = squaredDistance(query, node.cell.centerXY)
         if (distSq <= radiusSq) {
-            result.add(HexCellWithDistance(node.cell, sqrt(distSq)))
+            result.add(HexCellWithDistance(node.cell, sqrt(distSq).toDouble()))
         }
         val axis = node.axis
         val queryVal = if (axis == 0) query.x else query.y
@@ -142,7 +142,7 @@ class KDTree(
     private fun squaredDistance(
         a: Offset,
         b: Offset,
-    ): Double {
+    ): Float {
         val dx = a.x - b.x
         val dy = a.y - b.y
         return dx * dx + dy * dy
@@ -151,5 +151,5 @@ class KDTree(
     private fun distanceMeters(
         a: Offset,
         b: Offset,
-    ): Double = sqrt(squaredDistance(a, b))
+    ): Float = sqrt(squaredDistance(a, b))
 }
